LayoutTests/ietestcenter/Javascript/15.3.4.5-8-1-expected.txt

15.3.4.5-8-1 (TestCases/chapter15/15.3/15.3.4/15.3.4.5/15.3.4.5-8-1.js) Function.prototype.bind, type of bound function must be 'function'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL ES5Harness.preconditionPassed should be true (of type boolean). Was undefined (of type undefined).
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
